pyproject.toml
[project]
name = "openai-api-agent-project"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.3",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "json5>=0.12.1",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "yfinance>=0.2.65",
]
